requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.1
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
simpy==4.1.1
six==1.17.0
tzdata==2025.2
zipp==3.23.0



